# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReconFlow AI",
    "description": "Delivery-platform reconciliation for restaurants.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/uploaded-files/{uploaded_file_id}/validation": {
      "post": {
        "tags": [
          "data-validation"
        ],
        "summary": "Validate Uploaded File",
        "operationId": "validate_uploaded_file_uploaded_files__uploaded_file_id__validation_post",
        "parameters": [
          {
            "name": "uploaded_file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uploaded File Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FileValidationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "uploaded_file_id": {
            "type": "string",
            "title": "Uploaded File Id"
          },
          "status": {
            "$ref": "#/components/schemas/ValidationStatus"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "id",
          "uploaded_file_id",
          "status",
          "row_count",
          "checked_at",
          "issues"
        ],
        "title": "FileValidationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IssueCode": {
        "type": "string",
        "enum": [
          "MISSING_COLUMN",
          "NO_DATA_ROWS",
          "UNREADABLE_FILE",
          "TOO_MANY_ROWS",
          "INVALID_NUMBER",
          "INVALID_DATE",
          "EMPTY_REQUIRED_FIELD"
        ],
        "title": "IssueCode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationIssueResponse": {
        "properties": {
          "row_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Number"
          },
          "column_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column Name"
          },
          "issue_code": {
            "$ref": "#/components/schemas/IssueCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "row_number",
          "column_name",
          "issue_code",
          "message"
        ],
        "title": "ValidationIssueResponse"
      },
      "ValidationStatus": {
        "type": "string",
        "enum": [
          "PASSED",
          "FAILED"
        ],
        "title": "ValidationStatus"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}